Fail IOS cli init if "enable" couldnt be perfomed
If an IOS device returns something unexpected during "enable" phase
the unit would go through without an issue, however if a device is under heavy load
and it just returns "Pass" instead of the full message, the session becomes corrupted.

So rather than that, fail and a subsequent reconnect will be probably more lucky.

Change-Id: I355f516648a420e5f85d52fda51f35bf2b486717
Signed-off-by: Maros Marsalek <mmarsalek@frinx.io>